backend/requirements.txt
Django
Pillow
djangorestframework
django-filter
django-cors-headers
drf-yasg
gunicorn
psycopg2-binary
djangorestframework-simplejwt

